captainsmode
Added Suspend Mode processing to mod to allow it to be enabled but not
active.
Updated config AutoDisableSelf to "suspend" the mod instead of disabling
it.
Suspend action for AutoDisableSelf will occur on Post Map Load.

Updated messages to use the Locale message translation.
Removed config option for AutoDisableVoteRandom.
Added process to Override VoteRandom.ShuffleTeams. Override works when
mod is not suspended.
Added option to allow Votes for Randomize ReadyRoom and AddCommanderBots
when player count is lower than 12 (Config.BlockedVotesMinPlayer).

New command sh_cm_captainsnight "!captainsnight" activates team join
blocking and starts messaging for becoming a captain.
New command sh_cm_autodisable - makes it easier to set the
"AutoDisableSelf" config. AutoDisableSelf allows the mod to suspend
itself after a new map is loaded.

Updated to use DMD ForceBotDelay(86400) to help remove bots for the
Config.AutoRemoveBots option.

Changed when the dice is rolled for which captain gets which team. Now
its run when Captains window is started.
Added ability to save the last captain team names and return them when
requested.

votecaptainsmode
New extension to allow players to vote to activate the Captains mode.
This works with the new Suspend state.

rematch
New extension to allow for a Rematch that swaps the players to the other
team and starts a new game.
Will block VoteRandom and VoteMap.

voterematch
New extension to allow players to vote for a rematch.